+++ CHANGELOG.md
@@ -13,6 +13,7 @@ These notes are publicly posted in [production](https://wordplay.dev/updates), s
 ### Fixed
 
 - The to front and back buttons in the character editor didn't work.
+- Repaired a translation bug that prevented rendering of translated formatted text.
 
 ## 0.17.9 - 2026-05-09